changed create_point dtype

+++ src/artemis/utils.py
@@ -10,9 +10,9 @@ def create_point(*args):
             args[index] = 0
 
     if len(args) == 2:
-        return np.array([args[0], args[1]], dtype=np.int8)
+        return np.array([args[0], args[1]], dtype=np.float16)
     elif len(args) == 3:
-        return np.array([args[0], args[1], args[2]], dtype=np.int8)
+        return np.array([args[0], args[1], args[2]], dtype=np.float16)
     else:
         raise TypeError("Invalid number of arguments")
